Revert "move deprecated flags code to deprecated.go, refactor to allow migration of other flags"
This reverts commit fa393efa1d635660e5b2cc763f983a3614e8bdc3.
